:recycle: (Tests) Rename .test in .native and create .web only
